clean up the code a bit, add initial generator tests
